exam-check-tool-revamp/backend/requirements.txt
langchain_core==0.3.43
langchain_google_genai==2.0.11
langchain_groq==0.2.5
langchain_ollama==0.2.3
langchain_openai==0.3.8
langgraph==0.3.5
mysql_connector_repackaged==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
typing_extensions==4.12.2
Flask==3.1.0
flask_cors==5.0.1
flask_sqlalchemy==3.1.1
PyMySQL==1.1.1
